refactor: move webfinger + instance-discovery into discovery/ (Plan E)
