+++ src/discovery/dynamic-instance-discovery.ts
@@ -14,9 +14,9 @@ import {
   MAX_DYNAMIC_INSTANCES,
   REQUEST_TIMEOUT,
   USER_AGENT,
-} from "./config.js";
-import { LRUCache } from "./utils/lru-cache.js";
-import { validateExternalUrl } from "./utils.js";
+} from "../config.js";
+import { LRUCache } from "../utils/lru-cache.js";
+import { validateExternalUrl } from "../utils.js";
 
 const logger = getLogger("activitypub-mcp:discovery");
