+++ src/discovery/instance-discovery.ts
@@ -2,10 +2,10 @@ import { readFileSync } from "node:fs";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
 import { getLogger } from "@logtape/logtape";
-import { MAX_RESPONSE_SIZE, REQUEST_TIMEOUT, USER_AGENT } from "./config.js";
-import { readJsonWithLimit } from "./utils/fetch-helpers.js";
-import { validateExternalUrl } from "./utils.js";
-import { DomainSchema } from "./validation/schemas.js";
+import { MAX_RESPONSE_SIZE, REQUEST_TIMEOUT, USER_AGENT } from "../config.js";
+import { readJsonWithLimit } from "../utils/fetch-helpers.js";
+import { validateExternalUrl } from "../utils.js";
+import { DomainSchema } from "../validation/schemas.js";
 
 const logger = getLogger("activitypub-mcp");
