+++ src/mcp/tools.ts
@@ -9,9 +9,9 @@ import { getLogger } from "@logtape/logtape";
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { ErrorCode, McpError } from "@modelcontextprotocol/sdk/types.js";
 import { z } from "zod";
-import { dynamicInstanceDiscovery } from "../dynamic-instance-discovery.js";
+import { dynamicInstanceDiscovery } from "../discovery/dynamic-instance-discovery.js";
+import { instanceDiscovery } from "../discovery/instance-discovery.js";
 import { healthChecker } from "../health-check.js";
-import { instanceDiscovery } from "../instance-discovery.js";
 import { performanceMonitor } from "../performance-monitor.js";
 import { remoteClient } from "../remote-client.js";
 import { validateActorIdentifier, validateDomain, validateQuery } from "../server/index.js";
