+++ tests/integration/new-features-integration.test.ts
@@ -12,7 +12,7 @@
 
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { AuditLogger } from "../../src/audit/logger.js";
-import { DynamicInstanceDiscoveryService } from "../../src/dynamic-instance-discovery.js";
+import { DynamicInstanceDiscoveryService } from "../../src/discovery/dynamic-instance-discovery.js";
 import { InstanceBlocklist } from "../../src/instance-blocklist.js";
 
 describe.skipIf(!process.env.RUN_INTEGRATION_TESTS)("Audit Logger with Live Operations", () => {
